feat: add that like/dislike toggle
User can only like or dislike a post. Like the post again will toggle
the like. Dislike the post again will also toggle the dislike.

If user like the post, but now they decide to dislike, the like is
removed and the dislike is added.

If user dislike the post, but now they decide to like, the dislike is
removed and the like is added.